feat(mac): add CSMA/CA with backoff
Binary exponential backoff, CCA check before TX.
POSIX RSSI set to -90 dBm (below CCA threshold) for testing.
Refs: docs/IMPLEMENTATION.md krok 10

+++ src/mac/frame.c
@@ -19,7 +19,7 @@
  */
 
 umesh_result_t frame_serialize(const umesh_frame_t *frame,
-                               uint8_t *buf, uint8_t buf_size,
+                               uint8_t *buf, uint16_t buf_size,
                                uint8_t *out_len)
 {
     uint8_t total;
